Refactor: - clear separation between LinearProbing und Linked question - smaller function in multiple files
